IsNotEmpty failed to work with empty strings
Fixed by changing the condition to "argument.Length == 0". A null check is not needed as its only reference is to IsNotNullAndNotEmpty which checks for null before the empty. It is also faster than string.IsNullOrEmpty if the null check has already been done as shown in this StackOverflow link: https://stackoverflow.com/questions/10360370/why-is-string-isnullorempty-faster-than-string-length